Credit Risk Modeling
## About the Assignment (Project)

This dataset contains columns simulating credit bureau data. Credit default risk is the risk that a lender takes the chance that a borrower fails to make required payments of the loan. 

The main purpose of this analysis is to predict whether a new customer can be a reliable customer (Deciding on who to accept/refuse for a loan). It's a way to avoid default and increase the bank’s revenue. This can be used to automate approving and declining loan applications more accurately. 

## Approach

### To us a Logistic Regression Model
The dependent variable is Binary
The factor level 1 of the dependent variable will represent the desired outcome (default).
Only the meaningful variables will be included.
The independent variables should  be independent of each other. That is, the model should have little or no multicollinearity.
The independent variables are linearly related to the log odds.

## Feature description
### Feature Name ------------Description 
1. person_age ----------------Age 
2. person_income -------------Annual Income 
3. personhomeownership--------Home ownership 
4. personemplength------------Employment length (in years) 
5. loan_intent----------------Loan intent 
6. loan_grade-----------------Loan grade 
7. loan_amnt------------------Loan amount 
8. loanintrate----------------Interest rate 
9. loan_status----------------Loan status (0 is non default 1 is default) 
10. loanpercentincome---------Percent income 
11. cbpersondefaultonfile-----Historical default 
12. cbpresoncredhistlength----Credit history length